+++ pylibcklb/metadata.py
@@ -122,7 +122,7 @@ class Variables:
     The list trove classifiers applicable to this project
     """
 
-    install_requires = ["lxml", "markdown2", "numpy", "pyinstaller"]
+    install_requires = ["lxml", "markdown2", "numpy", "pyinstaller", "colorama"]
     """
     Python Packaging Index dependencies
     """